requirements.txt
qfieldcloud-sdk==0.5.1
python-dotenv==0.20.0
